+++ react/organizations/hooks.ts
@@ -2,7 +2,16 @@ import { useQuery } from 'react-apollo'
 
 import GET_ORGANIZATIONS from '../graphql/getOrganizations.graphql'
 
-export const INITIAL_FETCH_LIST_OPTIONS = {
+export interface FetchListOptions {
+  status: string[]
+  search: string
+  page: number
+  pageSize: number
+  sortOrder: string
+  sortedBy: string
+}
+
+export const INITIAL_FETCH_LIST_OPTIONS: FetchListOptions = {
   status: ['active', 'on-hold', 'inactive'],
   search: '',
   page: 1,
